renamed to run-commands-on-input
